requirements.txt
asgiref==3.7.2
attrs==23.1.0
boto3==1.28.49
botocore==1.31.49
Django==4.2.5
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
parameterized==0.9.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
s3transfer==0.6.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.16
